+++ src/main/java/insane96mcp/insanesurvivaloverhaul/data/generator/ISOBlockTagsProvider.java
@@ -19,6 +19,7 @@ import java.util.concurrent.CompletableFuture;
 
 public class ISOBlockTagsProvider extends BlockTagsProvider {
     public static final TagKey<Block> OBSIDIANS = create("obsidians");
+    public static final TagKey<Block> SKULLS = create("skulls");
     public static final TagKey<Block> INFESTED = create("infested");
     //public static final TagKey<Block> GRASS_BLOCKS = create("grass_blocks");
     //public static final TagKey<Block> TALL_GRASS = create("tall_grass");
@@ -52,6 +53,8 @@ public class ISOBlockTagsProvider extends BlockTagsProvider {
                 .add(Blocks.OBSIDIAN, Blocks.CRYING_OBSIDIAN);
         tag(INFESTED)
                 .add(Blocks.INFESTED_DEEPSLATE, Blocks.INFESTED_STONE, Blocks.INFESTED_COBBLESTONE, Blocks.INFESTED_STONE_BRICKS, Blocks.INFESTED_CRACKED_STONE_BRICKS, Blocks.INFESTED_CHISELED_STONE_BRICKS, Blocks.INFESTED_MOSSY_STONE_BRICKS);
+        tag(SKULLS)
+                .add(Blocks.SKELETON_SKULL, Blocks.SKELETON_WALL_SKULL, Blocks.WITHER_SKELETON_SKULL, Blocks.WITHER_SKELETON_WALL_SKULL, Blocks.CREEPER_HEAD, Blocks.CREEPER_WALL_HEAD, Blocks.DRAGON_HEAD, Blocks.DRAGON_WALL_HEAD, Blocks.PIGLIN_HEAD, Blocks.PIGLIN_WALL_HEAD, Blocks.PLAYER_HEAD, Blocks.PLAYER_WALL_HEAD, Blocks.ZOMBIE_HEAD, Blocks.ZOMBIE_WALL_HEAD);
 
         //noinspection unchecked
         tag(Tweaks.BREAK_ON_FALL)