+++ changelog.md
@@ -1,7 +1,7 @@
 ## Upcoming
 * Added back Mobs features: Equipment, Misc Mobs, Spawning, Zombie Siege
   * Not sure if and when Villager feature will come back.
-* Added back Movement features: Boats, Better climbable
+* Added back Movement features: Boats, Better climbable, Tagging, Swimming
   * The feature (that wasn't configurable) to make boats easier to break has been moved to Tweaks
   * Better climbing features no longer disable themselves if quark is present
 * Added back Misc features: Packs, Nerfs, Tweaks, Potions and Effects, Low Fish